Implement message controller.
This patchset contributes a message controller for Marconi client. It
implements simple property querying for messages, as well as the
ability to reload these properties as needed (polling?) and delete a
message. It also uses the assumed structure of the HTTPclient that
will form the basis of the connection from the common libraries.

six is used to boost py3k compat.

This also introduces a directory structure for supporting multiple
transports. Instead of having a 'controller' dir, now we have a
'transports/http' and 'transports/zmq'.

Author: Alejandro Cabrera
Implements: blueprint messages-management
Change-Id: I9d4e7c28939cdb68b74690e5c56e69d58ac1559a