Revert "Add apiclient library"
This reverts commit 4b9b0139641f0e85805783d69dfc3650885cd24a.

Rationale: the apiclient library needs many fixes and updates that we
are not able to address at this time. In the future, we may consider
returning to using it, but for now, we will implement a base
connection module from which this client will be able to speak to
marconi servers.

Change-Id: I40b053e5487348f524753db517b79bf4ca81d569