Implement queue's API methods
This patch implements methods for every queue operation. The 2
operations left out in this patch are: queue_list and queue_stats.

This patch makes the client, officially, usable as a library. An example
of how this could be use:

    client = Client()
    queue = client.queue('my_queue') # Creates the queue automatically
    queue.metadata() # Get's metadata from the queue
    queue.metadata(dict(new_meta='Yo yo!'))
    queue.delete()

Metadata was implemented as method instead of as a property to make it
explicit that both get and set do something in the remote server.

The client API was implemented in 2 separate levels. The highest level
is the one shown in the example above, which allows users to just query
Marconi instances without worrying about the transport, request build
process, references, deserealization and what not. The lowest instead,
allows the user to control every bit of the communication process - the
user can pick a specific transport, or pass custom params. For example -
which will make it possible to have other implementations around that
API in addition to the asynchronous support. Example of the API:

    transport = http.HttpTransport(cfg.CONF)
    request = request.prepare_request(cfg.CONF,
                                      endpoint='http://localhost:8888')
    core.queue_create(transport, request, 1, callback=my_callback)

There are tons of things to do and improve. For example:

    * The way config params are registered may work when using
    marconiclient as a library but they won't when using it as a CLI
    tool.
    * This code lacks of logging.
    * Handling of 20(1|4) is missing.

Partially-Implements blueprint python-marconiclient-v1
Implements blueprint queues-management

Change-Id: I8bdc8a4aff8ea22b5673bc7440e07796ecaf34cc