Bootstrap Messages support
This patch implements some of the operations required to add Message
support to marconiclient. The higher level implementation of message
deletion is not present in this patch, since it requires the definition
of a new Message type, which will submitted in the upcoming patch.

Partially-Implements blueprint python-marconiclient-v1
Partially-Implements blueprint messages-management

Change-Id: I42efe24e0e6083c530d494e95b66c6c01535a849