Make pbr only a build-time dependency.
This lets you build python-marconiclient packages that don't require pbr
to be installed at all. You would need pbr on the machine running
rpmbuild / debuild, but not on the machines that install the packages.

Unfortunately, this does not make python-marconiclient able to be
installed via pip 0.3.1 on Lucid; you'll need to uninstall the system
python-pip package and install a new pip some other way. Given that pip
< 1.3 doesn't perform SSL certificate validation for pypi (trivial MITM
attack, anyone?), you'd probably want to get a new pip anyway.

Change-Id: I2fb160b6dd446f1fc72eed2371db1afcb00b9e52