Add shard support to the client
The patch adds Shard support of the client. This will allow users to
create and delete shards from Marconi. The API requires the Marconi
instance to be and admin instance. However, there are no checks,
restrictions, yet.

This patch doesn't implement Shards listing. As soon as the
MessageIterator lands, it'll be possible to reuse that code for the
ShardsIterator.

Partially-Implements blueprint: shard-management
Partially-Implements blueprint: python-marconiclient-v1

Change-Id: I932c98d68a4a81b710dbea75a6bac123b7afaa08