Initial support for openstack-client
This patch adds the initial support for the common openstack client. In
order to use it, it is necessary to:

1. Install the common openstack client
2. Install Marconi's client library

Some examples:

    $  openstack queue list --limit 3
    $  openstack queue create myqueue
    $  openstack queue delete myqueue

The above assumes keystone is installed and has an endpoint configured
for marconi. In order to bypass this in a non-keystone based deployment,
the user can pass Marconi's url manually:

    $  openstack --os-token test --os-url http://127.0.0.1:8888 queue list

Environment variables are also supported, pls refer to the openstack
common client documentation.

Change-Id: I262e42a73e46149cb16e32f9a25b787433f6790e