Add support for `pop` implemented in v1.1
This patch implements pop in the client library. Pop is a new operation
that has been implemented in v1.1. It allows a client to remove N
messages from the top of queue. Although pop is a query parameter in the
server, it has been implemented as if it were an action.

Partially-Implements blueprint: api-v1.1

Change-Id: I29b771a6167185033821139bd795311f1a25a5e1