Add support for flavor to the client library
This patch adds support for flavors to the client library. Flavors are
an admin endpoint, which means pool's rules apply to them.

Since flavors were added since 1.1, this patch also adds a version
decorator that will help making sure features are accessible just if the
api_version the client has been instantiated with matches the version
the feature was added into.

Note that the core module does not use the version decorator. There are
2 reasons for this:

    1. Users accessing the core module should know what they're doing
    2. If the transport loads the spec for the API version it was
    instantiated with. If that spec does not contain a specific action,
    the whole operation will fail before hitting the wire.

Partially-Implements blueprint: api-v1.1

Change-Id: I92dbe417655a9eeefb030481a81a67840a63dd0e