Add docs for `Client` instances
This patch documents the basic usage of `Client` instances. It does not
provide a detailed documentation about the methods exposed but a high
level documentation of what the recommended way to access these
resources is.

The current version refers to sections like 'core API', 'config' and
others that haven't been documented yet. Once they will be, proper links
to those sections will be added here.

Change-Id: I705498b8e4c19c75c9697419d1fd90c9a9ae40b9