Gets 'pool' data if the resource exists
Pool instances can be created by just passing the name. They both
check at construction time whether the resource exist in the server or
not. If the resource doesn't exist, it'll attempt to create one.

Zaqarclient does not call anymore the 'pool_create' directly. It calls
'pool_get' first, which ensures the resource exists. if not, a 'pool'
will be created.

Change-Id: I8f18b5031c80cfd5744ba3b0b7058654934f44b4
Partial-Bug: #1396063