Gets 'flavor' data if the resource exists
Flavor instances can be created by just passing the name. it checks at
construction time whether the resource exist in the server or not. If
the resource doesn't exist, it'll attempt to create one.

Zaqarclient does not call anymore the 'flavor_create' directly. It
calls 'flavor_get' first, which ensures the resource exists. if not, a
'flavor' will be created.

Change-Id: Id78c79b7ead339cc64d57028eb49040fb67eb597
Closes-Bug: #1396063
Closes-Bug: #1394733