Added support to v2
Zaqarclient v2 currently has missing features that involve the
management of subscriptions, but in order to these to work it is
necessary to add other features such as 'queues', 'messages',
'claims', 'pools' and 'flavors'. I added 'queues' and 'messages' as a
priorital step before start working with the other features and with
subscriptions as well.

Change-Id: Iea082cbf9e1d0e4a745bc10e41e1206af6132fad