Fix freeze of Claim object's age property
The age property of Claim object is defined in a way that it only
queries claim's age from the server on the first call. On subsequent
calls it will just return the same age as in the first call.
Normally age property must return an updated age on each call, because
claim's age is something that changes with the time flow.

This patch makes Claim object's age property to return an updated value
gathered from the Zaqar server on each call.

Change-Id: Ifc3bd04242a6e4827e85f2be0c8bc6776b9e44ca
Closes-Bug: 1525994