Update pool on pool create if it exists
As of now on PUT, zaqar server updates pool if
it is already exists else it will create a new one.
However zaqar client return previous pool info if
it is already exists. This result to a confusion for
operator.

The zaqar client should maitain symmetry with zaqar server.

Change-Id: I0e7311321630c3f931cf27b0828394589eaf7320
Partial-Bug: #1532776