Fix queue stats for v1.1 and above versions
Now in the cli of queue status v1, it's using queue.exists() to
check the existance. However, the queue.exists isn't supported
in v1.1 and above. But the queue stats should be support for
all the versions. As a result, user will see errors like
'Unavailable on versions >= 1.1'. This patches fixes it.

Closes-Bug: #1534427

Change-Id: Iaf738af4eca2419c6eba0a410b1f7e45c9720f54