Create a websocket transport
This patch adds a new WebsocketTransport class that will help
consolidate common websocket client code that is useful
for connecting to and subscribing to Zaqar queues via
websockets.

We are already using a similar client within TripleO and
it makes sense to have it live in zaqarclient where it
can be shared and used by multiple projects.

Change-Id: Ia2a8deb599252d8308e44d595eb2bf443999aaad